Moved the method for doing manual file cleanup into a static class so it would be available everywhere.
